cmdline add option "-info diffFile"

+++ builds/android_ndk_jni_mk/hpatch.h
@@ -4,6 +4,7 @@
 #define hpatch_h
 #include <assert.h>
 #include <string.h>
+#include <stdint.h> //int64_t
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -17,7 +18,7 @@ extern "C" {
 
     // return THPatchResult, 0 is ok
     int hpatchz(const char *oldFileName,const char *diffFileName,
-                const char *outNewFileName, size_t cacheMemory) H_PATCH_EXPORT;
+                const char *outNewFileName,int64_t cacheMemory) H_PATCH_EXPORT;
 
 #ifdef __cplusplus
 }
